+++ GeneralsMD/Code/GameEngine/Source/GameLogic/Object/Collide/CrateCollide/SabotageInternetCenterCrateCollide.cpp
@@ -34,6 +34,7 @@
 
 // INCLUDES ///////////////////////////////////////////////////////////////////////////////////////
 #include "PreRTS.h"	// This must go first in EVERY cpp file int the GameEngine
+#include <cstdint>
 
 #include "Common/GameAudio.h"
 #include "Common/MiscAudio.h"
@@ -115,7 +116,7 @@ Bool SabotageInternetCenterCrateCollide::isValidToExecute( const Object *other )
 
 static void disableHacker( Object *obj, void *userData )
 {
-	UnsignedInt frame = (UnsignedInt)userData;
+	UnsignedInt frame = (UnsignedInt)(uintptr_t)userData;
 	if( obj )
 	{
 		obj->setDisabledUntil( DISABLED_HACKED, frame );
@@ -126,7 +127,7 @@ static void disableInternetCenterSpyVision( Object *obj, void *userData )
 {
 	if( obj && obj->isKindOf( KINDOF_FS_INTERNET_CENTER ) )
 	{
-		UnsignedInt frame = (UnsignedInt)userData;
+		UnsignedInt frame = (UnsignedInt)(uintptr_t)userData;
 
 		//Loop through all it's SpyVisionUpdates() and wake them all up so they can be shut down. This is weird because
 		//it's one of the few update modules that is actually properly sleepified.
